Migrate File Descriptor cmds to fs module
