+++ src/agent/fs.js
@@ -1,16 +1,10 @@
 'use strict';
 
-const { normalize } = require('path');
-const { platform, pointerSize } = Process;
 const { toByteArray } = require('base64-js');
-
-module.exports = {
-  ls,
-  cat,
-  open,
-  transformVirtualPath,
-  exist
-};
+const { normalize } = require('path');
+const path = require('path');
+const { platform, pointerSize } = Process;
+const sys = require('./sys');
 
 function debase (a) {
   if (a.startsWith('base64:')) {
@@ -635,3 +629,70 @@ function encodeBuf (buf, size, encoding) {
 
   return result.join('');
 }
+
+function listFileDescriptors (args) {
+  return listFileDescriptorsJson(args).map(([fd, name]) => {
+    return fd + ' ' + name;
+  }).join('\n');
+}
+
+function listFileDescriptorsJson (args) {
+  const PATH_MAX = 4096;
+  function getFdName (fd) {
+    if (sys._readlink && Process.platform === 'linux') {
+      const fdPath = path.join('proc', '' + sys.getPid(), 'fd', '' + fd);
+      const buffer = Memory.alloc(PATH_MAX);
+      const source = Memory.alloc(PATH_MAX);
+      source.writeUtf8String(fdPath);
+      buffer.writeUtf8String('');
+      if (sys._readlink(source, buffer, PATH_MAX) !== -1) {
+        return buffer.readUtf8String();
+      }
+      return undefined;
+    }
+    try {
+      // TODO: port this to iOS
+      const F_GETPATH = 50; // on macOS
+      const buffer = Memory.alloc(PATH_MAX);
+      const addr = Module.getExportByName(null, 'fcntl');
+      const fcntl = new NativeFunction(addr, 'int', ['int', 'int', 'pointer']);
+      fcntl(fd, F_GETPATH, buffer);
+      return buffer.readCString();
+    } catch (e) {
+      return '';
+    }
+  }
+  if (args.length === 0) {
+    const statBuf = Memory.alloc(128);
+    const fds = [];
+    for (let i = 0; i < 1024; i++) {
+      if (sys._fstat(i, statBuf) === 0) {
+        fds.push(i);
+      }
+    }
+    return fds.map((fd) => {
+      return [fd, getFdName(fd)];
+    });
+  } else {
+    const rc = sys._dup2(+args[0], +args[1]);
+    return rc;
+  }
+}
+
+function closeFileDescriptors (args) {
+  if (args.length === 0) {
+    return 'Please, provide a file descriptor';
+  }
+  return sys._close(+args[0]);
+}
+
+module.exports = {
+  listFileDescriptors,
+  listFileDescriptorsJson,
+  closeFileDescriptors,
+  ls,
+  cat,
+  open,
+  transformVirtualPath,
+  exist
+};
